# Speed up CPU sorting