@jagrit06 2024-03-19
@nicolov Thanks for the PR!

We chose `std::stable_sort` to be consistent in results with the GPU (which does a stable sort) and also to make sure that the order of elements in the sort forward is the same as the order we consider in sort backward (the sort vip uses an `argsort` which also does a stable sort). As such, going to `std::sort` without also producing / saving the indices at the same time like Pytorch does would make getting the correct vjp difficult