@awni 2024-03-20
Given @jagrit06 comments about the VJP/JVP it seems like it will be difficult to switch to this. Still I am somewhat surprised the stable sort is so much slower (4x!)... I wonder if we could write a faster one 🤔 .. 

@nicolov up to you if you want to investigate further or close this for now.